Day 53 Of C Programming
Write a Program To Take an Array Of Integers As Input, Calculate The Pivot Index Of This Array.
The Pivot Index Is The Index Where The Sum Of All The Numbers Strictly To The Left Of The Index Is Equal To The Sum Of All The Numbers Strictly To The Index's Right.
If The Index Is On The Left Edge Of The Array, Then The Left Sum Is 0 Because There Are No Elements To The Left, This Also Applies To The Right Edge Of The Array.
Print The Leftmost Pivot Index, If No Such Index Exists Return Appropriate Message.