Day 56 Of C Programming
Write a Program To Take an Array Of Integers As Input.
The Task Is To Find The Next Greater Element For Each Element Of The Array In Order Of Their Appearance In The Array.
Next Greater Element Of an Element In The Array Is The Nearest Element On The Right Which Is Greater Than The Current Element.
If There Does Not Exist Next Greater Of Current Element, Then Next Greater Element For Current Element Is -1.