Day 57 Of C Programming
Write a Program To Take an Array Of Integers As Input, The Task Is To Find The Previous Greater Element For Each Element Of The Array In Order Of Their Appearance In The Array.
Previous Greater Element Of an Element In The Array Is The Nearest Element On The Left Which Is Greater Than The Current Element.
If There Does Not Exist Next Greater Of Current Element, Then Previous Greater Element For Current Element Is -1.