Day 61 Of C Programming
Write a Program To Take An Integer Array arr & An Integer K As Inputs.
The Task Is To Find The First Negative Integer In Each Subarray Of Size K Moving From Left To Right.
If No Negative Exists In A Window, Print "0" For That Window.
Print The Results Separated By Spaces As Output.